Reorder tests (#615)
