+++ test/extensions/Project.toml
@@ -1,0 +1,47 @@
+name = "ClapeyronExtensionTests"
+uuid = "73f074b4-2fd7-4260-8743-9b275b5873e6"
+
+[deps]
+Clapeyron = "7c7805af-46cc-48c9-995b-ed0ed2dc909a"
+CoolProp = "e084ae63-2819-5025-826e-f8e611a84251"
+EoSSuperancillaries = "c1bf003f-4e47-49d9-bdfd-5a4051db3c04"
+Glenn = "d531567c-0c1c-43c6-9163-287e3bec9c15"
+# JutulDarcy = "82210473-ab04-4dce-b31b-11573c4f8e0a"
+Metaheuristics = "bcdb8e00-2c21-11e9-3065-2b553b22f898"
+MultiComponentFlash = "35e5bd01-9722-4017-9deb-64a5d32478ff"
+PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
+
+[compat]
+Clapeyron = "0.6"
+CoolProp = "0.2"
+EoSSuperancillaries = "1.2.3"
+Glenn = "0.3, 0.4"
+# JutulDarcy = "0.2, 0.3"
+Metaheuristics = "2, 3"
+MultiComponentFlash = "1.1.13"
+PythonCall = "0.9"
+Symbolics = "7"
+Unitful = "^1.12"
+
+[sources]
+Clapeyron = {path = "../.."}
+
+[extras]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
+Clapeyron = "7c7805af-46cc-48c9-995b-ed0ed2dc909a"
+CoolProp = "e084ae63-2819-5025-826e-f8e611a84251"
+EoSSuperancillaries = "c1bf003f-4e47-49d9-bdfd-5a4051db3c04"
+Glenn = "d531567c-0c1c-43c6-9163-287e3bec9c15"
+# JutulDarcy = "82210473-ab04-4dce-b31b-11573c4f8e0a"
+Metaheuristics = "bcdb8e00-2c21-11e9-3065-2b553b22f898"
+MultiComponentFlash = "35e5bd01-9722-4017-9deb-64a5d32478ff"
+PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
+
+[targets]
+test = ["Aqua", "Clapeyron", "CoolProp", "EoSSuperancillaries", "Glenn", "Metaheuristics", "MultiComponentFlash", "PythonCall", "Symbolics", "Test", "Unitful"]
