+++ test/extensions/src/ClapeyronExtensionTests.jl
@@ -1,0 +1,3 @@
+module ClapeyronExtensionTests
+
+end
