+++ test/extensions/test/runtests.jl
@@ -1,0 +1,9 @@
+using MultiComponentFlash: MultiComponentFlash
+using Symbolics
+using EoSSuperancillaries
+using PythonCall
+using Glenn
+using Aqua
+
+include("../../utils.jl")
+include("extensions.jl")
