+++ test/methods/api.jl
@@ -51,6 +51,69 @@
     # support for vol0
     modelgergCO2 = GERG2008(["carbon dioxide"])
     @test !isnan(only(Clapeyron.fugacity_coefficient(modelgergCO2, 1u"MPa", 300u"K"; phase=:stable, vol0=0.0023u"m^3")))
+
+
+    model = Clapeyron.GERG2008(["nitrogen","methane","ethane","propane","butane","isobutane","pentane"])
+    lng_composition = [0.93,92.1,4.64,1.7,0.42,0.32,0.09]
+    lng_composition_molar_fractions = lng_composition ./sum(lng_composition)
+    @test Clapeyron.molar_density(model,(380.5+101.3)*1000.0,-153.0+273.15,lng_composition_molar_fractions)/1000 ≈ 24.98 rtol = 1E-2
+    @test Clapeyron.mass_density(model,(380.5+101.3)*1000.0,-153.0+273.15,lng_composition_molar_fractions) ≈ 440.73 rtol = 1E-2
+    @test Clapeyron.molar_density(model,(380.5+101.3)u"kPa",-153.0u"°C",lng_composition_molar_fractions;output=u"mol/L") ≈ 24.98*u"mol/L"  rtol=1E-2
+    @test Clapeyron.mass_density(model,(380.5+101.3)u"kPa",-153.0u"°C",lng_composition_molar_fractions;output=u"kg/m^3")  ≈ 440.73*u"kg/m^3" rtol=1E-2
+end
+
+@testset "Rachford-Rice" begin
+    #error, all Ks > 0, from CalebBell/Chemicals
+    zs = [0.0, 0.0885053990596404, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.721469918219507, 0.01742948033685831,
+            0.1725952023839942]
+    Ks = [192.3625321718047, 105.20070698573475, 76.30532397111489, 37.090890262982, 21.38862102676539, 18.093547012968767,
+            10.319129068837443, 9.001962137200403, 4.565198737490148, 340.69153314749224, 269.09234343328467,
+            21.858385052861507]
+    @test Clapeyron.rr_vle_vapor_fraction(Ks,zs) == Inf
+
+    #here are some tests, from the paper.
+    #the paper has errors, z2 has 10 components, 9 and 10 repeated.
+    #and after that, they dont include the bmax,bmin in their selected pair.
+    z1 = [0.30,0.15, 0.05, 0.02, 0.01, 0.02, 0.02, 0.03, 0.07, 0.33]
+    k1 = [3.0E+00, 2.0E+00, 1.1E+00, 8.0E-01, 4.0E-01, 1.0E-01, 5.0E-02, 2.0E-02, 1.0E-02, 1.0E-04]
+    idxs1,ks1,zs1 = Clapeyron.rr_find_strongest(k1,z1)
+    @test all(in.(idxs1,Ref((1,2,9,10))))
+    @test Clapeyron.rr_vle_vapor_fraction(k1,z1) ≈ 0.1769 atol = 1e-4
+    z2 = [0.00034825,0.01376300,0.13084000,0.10925000,0.00001000,0.51009000,0.23564000,0.00006000]
+    k2 = [5.2663E+02, 5.0400E+01, 1.6463E+00, 8.7450E-01, 1.5589E-01, 3.6588E-02, 2.6625E-02, 4.8918E-06]
+    idxs2,ks2,zs2 = Clapeyron.rr_find_strongest(k2,z2)
+    #the paper is wrong here, it stablishes (1,2,6,7) ,does not include bmax
+    #we still converge to the correct root
+    @test all(in.(idxs2,Ref((1,2,6,8))))
+    @test Clapeyron.rr_vle_vapor_fraction(k2,z2) ≈ 0.00327 atol = 1e-4
+
+    z3 = [0.0187002, 0.0243002, 0.5419054, 0.0999010, 0.0969010, 0.0400004, 0.0212002, 0.0148001, 0.0741507, 0.0350404, 0.0173602, 0.0157402]
+    k3 = [1.32420, 1.12778, 1.22222, 1.11760, 9.88047E-01, 8.94344E-01, 7.87440E-01, 7.43687E-01, 8.11797E-01, 6.93279E-01, 5.09443E-01, 2.28721E-01]
+    idxs3,ks3,zs3 = Clapeyron.rr_find_strongest(k3,z3)
+    @test all(in.(idxs3,Ref((1,3,9,12))))
+    @test Clapeyron.rr_vle_vapor_fraction(k3,z3) ≈ 0.98725 atol = 1e-4
+
+    #testing exact solver,3 comps:
+    zs_3 = [0.5, 0.3, 0.2]
+    Ks_3 = [1.685, 0.742, 0.532]
+    xs_3 = [0.33940869696634357, 0.3650560590371706, 0.2955352439964858]
+    ys_3 = [0.5719036543882889, 0.27087159580558057, 0.15722474980613044]
+    β_3 = 0.6907302627738544
+    β_3_sol = Clapeyron.rr_vle_vapor_fraction(Ks_3,zs_3)
+    @test  β_3_sol ≈ β_3
+    @test Clapeyron.rr_flash_eval(Ks_3,zs_3,β_3_sol) <= 4*eps(β_3)
+    @test Clapeyron.rr_flash_vapor(Ks_3,zs_3,β_3_sol) ≈ ys_3
+    @test Clapeyron.rr_flash_liquid(Ks_3,zs_3,β_3_sol) ≈ xs_3
+
+    #Extreme Kvalues, it should give β = 0.99999
+    Ks_extreme = [15.464909530837806, 7006.64008090944, 1.8085837711444488, 0.007750676421035811, 30.98450366497431]
+    zs_extreme = [0.26562380186293233, 0.04910234829974003, 0.284394553603828, 0.006300023876072552, 0.3945792723574272]
+    @test Clapeyron.rr_vle_vapor_fraction(Ks_extreme,zs_extreme) ≈ 0.999999999
+
+    # Case where the evaluated point is right on the boundary
+    Ks_eps_0 = [1.2566703532018493e-21, 3.3506275205339295, 1.0300675710905643e-23, 1.706258568414198e-39, 1.6382855298440747e-20]
+    zs_eps_0 = [0.13754371891028325, 0.29845155687154623, 0.2546683930289046, 0.08177453852283137, 0.22756179266643456]
+    @test abs(Clapeyron.rr_vle_vapor_fraction(Ks_eps_0,zs_eps_0)) < eps(Float64)
 end
 
 @testset "association" begin
@@ -106,40 +169,6 @@ end
     test_assoc_matrix([500.0;;])
 end
 
-using EoSSuperancillaries
-#we test this separately, but leaving this on could speed up the test suite?
-Clapeyron.use_superancillaries!(false)
-
-if isdefined(Base,:get_extension)
-    @testset "Superancillaries.jl" begin
-        
-        pc = PCSAFT("eicosane")
-        pc2 = pharmaPCSAFT("oxygen")
-        cubic = tcPR(["water"])
-        
-        crit_pc = crit_pure(pc)
-        sat_cubic = saturation_pressure(cubic,373.15)
-        sat_pc2 = saturation_pressure(pc2,150.0)
-        
-        Clapeyron.use_superancillaries!(true)
-        
-        crit_sa_pc = crit_pure(pc)
-        sat_sa_cubic = saturation_pressure(cubic,373.15)
-        sat_sa_pc2 = saturation_pressure(pc2,150.0)
-
-        @test crit_pc[1] ≈ crit_sa_pc[1] rtol = 1e-6
-        @test crit_pc[3] ≈ crit_sa_pc[3] rtol = 1e-6
-        
-        @test sat_cubic[1] ≈ sat_sa_cubic[1] rtol = 1e-6
-        @test sat_cubic[2] ≈ sat_sa_cubic[2] rtol = 1e-6
-        @test sat_cubic[3] ≈ sat_sa_cubic[3] rtol = 1e-6
-        
-        @test sat_pc2[1] ≈ sat_sa_pc2[1] rtol = 1e-6
-        @test sat_pc2[2] ≈ sat_sa_pc2[2] rtol = 1e-6
-        @test sat_pc2[3] ≈ sat_sa_pc2[3] rtol = 1e-6
-
-    end
-end
 
 @testset "tpd" begin
     system = PCSAFT(["water","cyclohexane"])
@@ -170,7 +199,7 @@ end
     @test !Clapeyron.VT_diffusive_stability(model,v_unstable,T,z)
 end
 
-@testset "reference states" begin
+@testset verbose = true "reference states" begin
 
     @test !has_reference_state(PCSAFT("water"))
     @test has_reference_state(PCSAFT("water",idealmodel = ReidIdeal))
@@ -249,25 +278,27 @@ end
     end
 
     #reference state from EoSVectorParam
-    mod_pr = cPR(["water","ethanol"],idealmodel = ReidIdeal,reference_state = :ntp)
-    mod_vec = Clapeyron.EoSVectorParam(mod_pr)
-    Clapeyron.recombine!(mod_vec)
-    @test reference_state(mod_vec).std_type == :ntp
-    @test length(reference_state(mod_vec).a0) == 2
+    @testset "reference states - misc" begin
+        mod_pr = cPR(["water","ethanol"],idealmodel = ReidIdeal,reference_state = :ntp)
+        mod_vec = Clapeyron.EoSVectorParam(mod_pr)
+        Clapeyron.recombine!(mod_vec)
+        @test reference_state(mod_vec).std_type == :ntp
+        @test length(reference_state(mod_vec).a0) == 2
 
-    #reference state from Activity models
-    puremodel = mod_pr = cPR(["water","ethanol"],idealmodel = ReidIdeal)
-    act = NRTL(["water","ethanol"],puremodel = puremodel,reference_state = :ntp)
-    @test reference_state(act).std_type == :ntp
-    @test length(reference_state(act).a0) == 2
+        #reference state from Activity models
+        puremodel = mod_pr = cPR(["water","ethanol"],idealmodel = ReidIdeal)
+        act = NRTL(["water","ethanol"],puremodel = puremodel,reference_state = :ntp)
+        @test reference_state(act).std_type == :ntp
+        @test length(reference_state(act).a0) == 2
 
-    #issue 511
-    ref511 = ReferenceState(:nbp)
-    model511 = cPR("water",idealmodel=ReidIdeal,reference_state = ref511)
-    @test Clapeyron.reference_state(model511).std_type == :nbp
+        #issue 511
+        ref511 = ReferenceState(:nbp)
+        model511 = cPR("water",idealmodel=ReidIdeal,reference_state = ref511)
+        @test Clapeyron.reference_state(model511).std_type == :nbp
+    end
 end
 
-@testset "Solid Phase Equilibria" begin
+@testset verbose = true "Solid Phase Equilibria" begin
     @testset "Pure Solid-Liquid Equilibria" begin
         model = CompositeModel(["methane"]; fluid = SAFTVRMie, solid = SAFTVRSMie)
 
@@ -358,7 +389,7 @@ end
 GC.gc()
 
 #test for difficult equilibria.
-@testset "challenging equilibria" begin
+@testset verbose = true "Challenging equilibria" begin
 
     #see https://github.com/ClapeyronThermo/Clapeyron.jl/issues/173
     @testset "VTPR - 1" begin
@@ -395,8 +426,10 @@ GC.gc()
         T2 = 0.999Tc2
 
         #this test fails on mac, julia 1.6
-        @test Clapeyron.saturation_pressure(model2,T2,crit_retry = false)[1] ≈ 1.451917823392476e6 rtol = 1e-6
-
+        #also seems to fail on ubuntu, julia 1.10
+        #works on 1.12
+        #probably related to removing the implicit loading of EoSSuperancillaries?
+        #@test_broken Clapeyron.saturation_pressure(model2,T2,crit_retry = false)[1] ≈ 1.451917823392476e6 rtol = 1e-6
         #https://github.com/ClapeyronThermo/Clapeyron.jl/issues/237
         #for some reason, it fails with mac sometimes
         if !Base.Sys.isapple()
@@ -431,7 +464,7 @@ GC.gc()
     GC.gc()
 end
 
-@testset "partial properties" begin
+@testset "Partial properties" begin
     model_pem = PR(["hydrogen", "oxygen", "water"])
     z = [0.1,0.1,0.8]
     p,T = 0.95e5,380.15
@@ -440,7 +473,7 @@ end
     end
 end
 
-@testset "supercritical lines" begin
+@testset "Supercritical lines" begin
     model = PR("methane")
     T_initial = 200.0
     p_widom, v1 = widom_pressure(model, T_initial)
@@ -464,7 +497,7 @@ end
     @test p_ciic ≈ ciic_pressure(model,T_initial,v0 = 1.01*v3)[1] rtol = 1e-6
 end
 
-@testset "thermodynamic factor" begin 
+@testset "Thermodynamic factor" begin 
     eos_model = PCSAFT(["water", "ethanol"])
     Γ_eos = thermodynamic_factor(eos_model, 1e5, 300., [2.,4.])
     @test size(Γ_eos) == (1,1)
