+++ test/methods/eos.jl
@@ -1,5 +1,3 @@
-using Clapeyron, Test, Unitful
-
 @testset "pharmaPCSAFT, single components" begin
     system = pharmaPCSAFT(["water"])
     v1 = Clapeyron.saturation_pressure(system, 280.15)[2]
@@ -425,14 +423,6 @@ end
 
 @testset "GERG2008 methods, multi-components" begin
     @testset "Bulk properties" begin
-        model = Clapeyron.GERG2008(["nitrogen","methane","ethane","propane","butane","isobutane","pentane"])
-        lng_composition = [0.93,92.1,4.64,1.7,0.42,0.32,0.09]
-        lng_composition_molar_fractions = lng_composition ./sum(lng_composition)
-        @test Clapeyron.molar_density(model,(380.5+101.3)*1000.0,-153.0+273.15,lng_composition_molar_fractions)/1000 ≈ 24.98 rtol = 1E-2
-        @test Clapeyron.mass_density(model,(380.5+101.3)*1000.0,-153.0+273.15,lng_composition_molar_fractions) ≈ 440.73 rtol = 1E-2
-        @test Clapeyron.molar_density(model,(380.5+101.3)u"kPa",-153.0u"°C",lng_composition_molar_fractions;output=u"mol/L") ≈ 24.98*u"mol/L"  rtol=1E-2
-        @test Clapeyron.mass_density(model,(380.5+101.3)u"kPa",-153.0u"°C",lng_composition_molar_fractions;output=u"kg/m^3")  ≈ 440.73*u"kg/m^3" rtol=1E-2
-    
         #test found in #371
         model2 = GERG2008(["carbon dioxide","nitrogen","water"])
         @test mass_density(model2,64.0e5,30+273.15,[0.4975080785711593, 0.0049838428576813995, 0.4975080785711593],phase = :l) ≈ 835.3971524715569 rtol = 1e-6
